video_to_slice/requirements.txt
# Google Video Intelligence批量视频切片工具依赖包
# 安装命令: pip install -r requirements.txt

# Google Cloud Video Intelligence API
google-cloud-videointelligence>=2.11.0

# Google Cloud Storage (用于大文件上传)
google-cloud-storage>=2.8.0

# 其他核心依赖
requests>=2.31.0

# 可选依赖 (如果需要的话)
# moviepy>=1.0.3  # 用于本地视频处理，可选

# ========================================
# 系统要求
# ========================================
# - Python 3.10+
# - FFmpeg (用于视频切片)
# - FFprobe (用于视频信息获取)
# 
# FFmpeg安装:
# Ubuntu/Debian: sudo apt install ffmpeg
# macOS: brew install ffmpeg  
# Windows: 下载并安装 https://ffmpeg.org/download.html
#
# ========================================
# Google Cloud配置
# ========================================
# 1. 创建Google Cloud项目
# 2. 启用Video Intelligence API
# 3. 创建服务账号密钥
# 4. 下载JSON凭据文件
# 5. 设置环境变量:
#    export GOOGLE_APPLICATION_CREDENTIALS="path/to/credentials.json"
#    或将凭据文件命名为 google_credentials.json 放在工具目录下
#
# ========================================
# 使用说明
# ========================================
# python batch_video_to_slice.py input_videos/ -o output_slices/ 